chore(accounts/service/fetch): rename to `FetchService`
